refactor(routes): artifacts 세션 접근 헬퍼 추출 — 파일 크기 가드(600줄) 준수
PR #252 머지로 artifacts.routes.ts 가 613줄이 되어 CI Gate 3 실패.
기존 artifact-viewer-service 추출과 동일 패턴으로 세션 접근 검증 헬퍼
(resolveUserId/resolveAnonSessionId/assertSessionAccess/sendSessionAccessError)를
routes/artifact-session-access.ts 로 이동 (동작 변경 없음, 613→567줄).
